+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * added `stubbles\streams\StandardInputStream`
   * added `stubbles\streams\StandardOutputStream`
   * annotated `stubbles\lang\errorhandler\ExceptionLogger` with `@Singleton`
+  * added `stubbles\lang\castToArray()`
 
 
 5.3.2 (2015-03-09)